requirements.txt
# SmartVoice Chat - Python Dependencies

# Core dependencies
sherpa-onnx>=1.10.0
PyYAML>=6.0

# Optional dependencies
numpy>=1.20.0
